move compute_boxplot_metrics to utils
